Relax check around unmount
Change zvol BUSY, query remove device

Instead of checking for ZVOL EBUSY against 1, which breaks
for destroying zvol, we skip the sync check in ZVOL create.

QUERY_REMOVE_DEVICE is called on mounts-in-progress, we
need to return failure, or it will not carry on with mount.

We should only unload the bus/module when we actually asked
for it, or it happens when Windows checks to unload a disk.

Update the EV codesign path to handle OpenZVOL.sys

The wait-for-mount started already signalled.

Also, that wasn't quite 10s.

Can't use system_taskq for long time threads

The wait for OpenZVOL can wait forever, and many things
call taskq_wait() to make sure their thread has finished.

Change to using a thread.

Loosen the locking in writes and update disk counters.

Attempt to detect if a disk is busy

Send GUID_IO_VOLUME_DISMOUNT in unmount

Attach devices in mount

Bubble up IRPs like we are supposed to

Assign Vpb so we dont leak references

Unmount mountmgr work should be on dcb

Make sure to get newPDO for mounts

Or Windows uses the last Vpb for queries, which
goes very wrong.

More mount/unmount corrections

Also fix check_disk checking for ZFS
partitions.
Fix printing of errors

+++ lib/libspl/include/sys/uio.h
@@ -87,7 +87,7 @@ typedef struct zfs_uio {
 static inline boolean_t
 zfs_dio_page_aligned(void *buf)
 {
-	return ((((unsigned long)(buf) & (PAGESIZE - 1)) == 0) ?
+	return ((((unsigned long long)(buf) & (PAGESIZE - 1)) == 0) ?
 	    B_TRUE : B_FALSE);
 }
